requirements/base.txt
# common dependencies used for all environments
Django==3.2.8
django-cors-headers==3.10.0
django-extensions==3.1.5
djangorestframework==3.12.4
django-notifications-hq==1.6.0
django-storages==1.12.3

whitenoise==5.3.0
psycopg2-binary==2.9.1

django-taggit==1.5.1
Pillow==8.4.0
requests==2.26.0